pyproject.toml
[tool.poetry]
name = "whisper"
version = "0.1.0"
description = "Whisper api"
authors = ["Valentin Lapparov <vlapparov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.3"
openai-whisper = "^20240930"
uvicorn = "^0.32.0"
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
